Working first implementation of Datahike JDBC backend
This is a first working version of Datahike JDBC backend which
builds on top of https://github.com/alekcz/konserve-jdbc. The
current version of konserve-jdbc builds on clojure.java.jdbc
which is superseded by next.jdbc. It seems that next.jdbc
is more future proof and this should be ported in the future.

Configuration is done by using a config map as parameter to
the corresponding functions or by using the appropriate
environment variables or system properties. This works like
other backends, but since JDBC is very versatile you can
only pass the whole configuration as one string to one
environment variable/system property named DATAHIKE_STORE_CONFIG

But you need to set DATAHIKE_STORE_BACKEND=jdbc as well that
datahike knows which backend to speak to. There are integration
tests that test the integration with Datahike in MySQL,
PostgreSQL and H2.

- Refers https://github.com/replikativ/datahike/issues/161
- Needs to be merged together with
  https://github.com/replikativ/datahike/pull/195