Add jdbcUrl as configuration parameter
next.jdbc can take urls via :jdbcUrl to configure your connection.
For that to work it was necessary to make dbtype optional as
configuration parameter and add jdbcUrl as optional configuration
parameter. Now it is possible to pass the configuration partially
as url and as key-value-pairs to configure the connection.

- Closes #8
- Updated README
- New CircleCI step for integrationtesting with postgresql and
  mariadb